Integración módulo básico de meteoroloxía (OpenWeather). Sesión 17

+++ app/lib/views/dashboard/widgets/weather_card.dart
@@ -8,7 +8,7 @@ class WeatherCard extends StatelessWidget {
     required this.location,
   });
 
-  final int temperature;
+  final double temperature;
   final String condition;
   final String location;
